fix: update tests to avoid to connect to mongodb
when using Document.deleteOne(), the tests were using the internal
implementation of mongoose to try to delete the document, making the
tests fail when no mongoose server was available. Furthermore,
mongoose.connect function was also trying to connect to the database,
what was fixed importing mongoose directly on jest.setup.js. Finally,
the findOneAndDelete pre hook was removed from the coverage, since
mockingoose doesn't mock it